myproject-core/pyproject.toml
[project]
name = "myproject-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nguyentran0212", email = "gen.tran.1991@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.81.7",
    "pydantic-settings>=2.12.0",
    "python-frontmatter>=1.1.0",
]

[build-system]
requires = ["uv_build>=0.9.27,<0.10.0"]
build-backend = "uv_build"
